traduje clasification y añadi las funciones
